+++ ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: jprofit <jprofit@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/14 18:16:51 by jprofit           #+#    #+#             */
-/*   Updated: 2022/11/17 15:43:16 by jprofit          ###   ########.fr       */
+/*   Updated: 2022/11/18 16:52:15 by jprofit          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (s != NULL)
